TECH-12593: separate Context from ContextHandler
