+++ spec/lib/contextual_logger/context_spec.rb
@@ -3,16 +3,16 @@
 require 'spec_helper'
 require 'contextual_logger'
 
+class ContextualLoggerContextSpecContainer
+  include ::ContextualLogger::Context
+end
+
 RSpec.describe ContextualLogger::Context do
   let(:context) { { service: { name: 'tts', description: 'TTS' }, integration: "google" } }
   let(:context2) { { service: { description: 'Context 2' }, integration: "google" } }
   let(:context3) { { service: { name: 'Context 3' } } }
 
-  class ContextualLoggerContextContainerForSpec
-    include ::ContextualLogger::Context
-  end
-
-  let(:instance) { ContextualLoggerContextContainerForSpec.new }
+  let(:instance) { ContextualLoggerContextSpecContainer.new }
 
   describe 'mixin methods' do
     describe 'current_context/current_context=' do
@@ -34,7 +34,7 @@ RSpec.describe ContextualLogger::Context do
 
       it 'the current_context= values are separately per containing instance' do
         instance.current_context = context
-        instance2 = ContextualLoggerContextContainerForSpec.new
+        instance2 = ContextualLoggerContextSpecContainer.new
         instance2.current_context = context2
 
         expect(instance.current_context({})).to eq(context)
@@ -46,19 +46,6 @@ RSpec.describe ContextualLogger::Context do
 
         expect(instance.current_context(context)).to eq(context)
       end
-    end
-  end
-
-  describe ContextualLogger::Context::Handler do
-    subject(:handler) { described_class.new(instance, context) }
-
-    it { is_expected.to respond_to(:reset!) }
-
-    it 'resets the thread context on reset!' do
-      instance.current_context = context2
-      handler.reset!
-
-      expect(instance.current_context(nil)).to eq(context)
     end
   end
 end
